imageGen/requirements.txt
fastapi==0.119.0
torch==2.9.0+cu128
torchvision==0.24.0+cu128
uvicorn[standard]==0.37.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
pydantic==2.12.0
python-multipart==0.0.20
python-dotenv==1.1.1
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
sentence-transformers==5.1.1
numpy==2.2.6
diffusers==0.35.2
transformers==4.57.1
opencv-python==4.12.0.88
accelerate==1.11.0
imageio==2.34.0
imageio-ffmpeg==0.5.1